Fix visualization imports,collapsed tests into one and update documentation
